test(dataset): split the 2261-line test_plot.py by concern (Theme E.1)
Break the 16-class plot file into five concern modules — test_plot_{dataset,color,facade,
engine,cleanups}.py. Each re-declares the cleopatra importorskip guard, the
Config.set_matplotlib_backend("agg") call (moved out of the TestPlotDataSet class body),
and the autouse figure-teardown fixture. _make_nc_subset_with_band_count stays with its sole
consumer (TestNetCDFPlot). Pure move: 115 tests collected before and after, all green under
Agg. No test logic changes.
